git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@723 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/src/qwt_abstract_scale_draw.h
@@ -72,8 +72,8 @@ public:
     void setSpacing(double margin);
     double spacing() const;
         
-    void setPenWidth(double width);
-    double penWidth() const;
+    void setPenWidth(int width);
+    int penWidth() const;
 
     virtual void draw(QPainter *, const QPalette &) const;
